Adds a handicap: Broken Finger
BF prevents the player from wielding a weapon. This is meant to
encourage playstyles like pure-mage, fist-fighting, and
illusion-control.

As of right now, the patch limits this handicap at highest priority to
khajit only, but ideally, I think it should apply to everybody but
with a priority boost for khajit only. Maybe.